+++ OnRail/Extensions/OnFail/OnFailExtensions.ThrowExceptionExtension.cs
@@ -1,9 +1,10 @@
+using OnRail.Extensions.ThrowException;
 using OnRail.ResultDetails;
 
-namespace OnRail.Extensions.ThrowException;
+namespace OnRail.Extensions.OnFail;
 
-public static partial class ThrowExceptionExtensions {
-    public static Result<T> ThrowExceptionOnFail<T>(this Result<T> source) {
+public static partial class OnFailExtensions {
+    public static Result<T> OnFailThrowException<T>(this Result<T> source) {
         if (!source.IsSuccess) {
             source.Detail ??= new ErrorDetail();
             source.Detail.ThrowException();
@@ -12,7 +13,7 @@ public static partial class ThrowExceptionExtensions {
         return source;
     }
 
-    public static Result ThrowExceptionOnFail(this Result source) {
+    public static Result OnFailThrowException(this Result source) {
         if (!source.IsSuccess) {
             source.Detail ??= new ErrorDetail();
             source.Detail.ThrowException();